FEAT authorization,policy on CRUD

+++ app/Http/Controllers/DishController.php
@@ -58,6 +58,9 @@ class DishController extends Controller
      */
     public function show(dish $dish)
     {
+        //$this->authorize('view', $dish);
+        //Se desidero oscurare anche la vista show del piatto riabilito la riga sopra ed riabilito la riga nella sezione view nel file DishPolicy
+
         return view('dishes.show', compact('dish'));
     }
 
@@ -81,6 +84,8 @@ class DishController extends Controller
      */
     public function update(UpdatedishRequest $request, dish $dish)
     {
+        $this->authorize('update', $dish);
+
         $data = $request->validated();
         $dish->update($data);
 
@@ -95,7 +100,7 @@ class DishController extends Controller
      */
     public function destroy(dish $dish)
     {
-        //$dish->delete();
+        $this->authorize('delete', $dish);
 
         if($dish->trashed()){
             $dish->forceDelete();
